adding scripts
